requirements.txt
flask==3.1.3
flask-bcrypt==1.0.1
gunicorn
torch
torchvision
Pillow
streamlit